@@ -1,3 +1,6 @@
+[tool.pdm.venv]
+prompt = "aoc"
+
 [tool.pdm.scripts]
 watch = "python -m aocli -w"
 
@@ -15,5 +18,5 @@ dependencies = [
     "watchdog>=3.0.0",
     "websockets>=12.0",
 ]
-requires-python = ">=3.11,<3.13"
+requires-python = ">=3.12,<3.13"
 license = { text = "MIT" }
